Refactor styles for navbar.
* Other general style changes from SCSS to CSS.
