+++ __tests__/components/AppNavbar.test.jsx
@@ -1,6 +1,6 @@
 import { expect, test } from "vitest";
 import { render, screen } from "@testing-library/react";
-import AppNavbar from "../../src/components/layout/AppNavbar";
+import AppNavbar from "@/components/layout/navbar/AppNavbar";
 
 render(<AppNavbar />);
