+++ src/app/-styles/main.css
@@ -1,0 +1,7 @@
+h1,
+h2,
+h3,
+h4,
+h5 {
+  text-wrap: pretty;
+}
