+++ src/app/-styles/typography.css
@@ -134,7 +134,7 @@ Application typography. The Inter font family is applied by next/font in layout.
   font-weight: 700;
   font-size: 48px;
   line-height: 56px;
-  letter-spacing: -4%;
+  letter-spacing: 0.04em;
 }
 
 /* The small font above each of the sections' header (refer to home page) */
@@ -155,7 +155,7 @@ Application typography. The Inter font family is applied by next/font in layout.
   font-weight: 700;
   font-size: 31px;
   line-height: 38px;
-  letter-spacing: -2%;
+  letter-spacing: 0.02em;
 }
 
 .h3-semi-bold {
