+++ src/app/global.css
@@ -1,38 +1,68 @@
-@import "../styles/main.css";
 @import "tailwindcss";
-
+@import "./-styles/main.css";
+@import "./-styles/typography.css";
+@import "./-styles/layout.css";
+@import "./-styles/button.css";
+@import "../components/layout/navbar/app-navbar.css";
+@import "../components/layout/navbar/extended-menu.css";
 @plugin "@tailwindcss/typography";
 
-/* This file is strictly for overriding the tailwind theme. Use styles/base/base.scss for global styles. */
+/* This file is strictly for overriding the tailwind theme. Use -styles/main.css for global styles. */
 @theme {
-  --color-neutral-50: #f7f7f8;
-  --color-neutral-100: #eaeaec;
-  --color-neutral-200: #d4d4d8;
-  --color-neutral-300: #bfbfc5;
-  --color-neutral-400: #a9aab1;
-  --color-neutral-500: #72727d;
-  --color-neutral-600: #55555e;
-  --color-neutral-700: #3a3a40;
-  --color-neutral-800: #1d1d20;
-  --color-neutral-900: #060717;
-  --color-primary-50: #f0f4fe;
-  --color-primary-100: #d6e1fe;
-  --color-primary-200: #adc4fc;
-  --color-primary-300: #85a6fb;
-  --color-primary-400: #5b89fa;
-  --color-primary-500: #0d2866;
-  --color-primary-600: #0034ae;
-  --color-primary-700: #002377;
-  --color-primary-800: #00113c;
-  --color-primary-900: #00091e;
-  --color-secondary-50: #fffcf5;
-  --color-secondary-100: #fff9e7;
-  --color-secondary-200: #fff2cf;
-  --color-secondary-300: #ffe5a1;
-  --color-secondary-400: #ffd874;
-  --color-secondary-500: #ffbf32;
-  --color-secondary-600: #ecaf2c;
-  --color-secondary-700: #e9ad2e; /* Check this one, very similar to 600 */
-  --color-secondary-800: #251e0d;
-  --color-secondary-900: #0c0900;
+  /* Screen breakpoints. */
+  --screen-xs: 400px;
+  --screen-sm: 576px;
+  --screen-md: 768px;
+  --screen-lg: 992px;
+  --screen-xl: 1200px;
+  --screen-xxl: 1400px;
+
+  --neutral-50: #f7f7f8;
+  --neutral-100: #eaeaec;
+  --neutral-200: #d4d4d8;
+  --neutral-300: #bfbfc5;
+  --neutral-400: #a9aab1;
+  --neutral-500: #72727d;
+  --neutral-600: #55555e;
+  --neutral-700: #3a3a40;
+  --neutral-800: #1d1d20;
+  --neutral-900: #060718;
+  --primary-50: #f0f4fe;
+  --primary-100: #d6e1fe;
+  --primary-200: #adc4fc;
+  --primary-300: #85a6fb;
+  --primary-400: #5b89fa;
+  --primary-500: #0d2866;
+  --primary-600: #0034ae;
+  --primary-700: #002377;
+  --primary-800: #00113c;
+  --primary-900: #00091e;
+  --secondary-50: #fffcf5;
+  --secondary-100: #fff9e7;
+  --secondary-200: #fff2cf;
+  --secondary-300: #ffe5a1;
+  --secondary-400: #ffd874;
+  --secondary-500: #ffbf32;
+  --secondary-600: #ecaf2c;
+  --secondary-700: #e9ad2e;
+  --secondary-800: #251e0d;
+  --secondary-900: #0c0900;
+  --shades-black: #000000;
+  --shades-white: #ffffff;
+
+  /* Brand colors. */
+  --landing-background-color: var(--secondary-50);
+  --navbar-color: var(--secondary-50);
+  --main-vivid-blue: #00478e;
+  --main-light-blue: #00a6fb;
+  --main-grey: #72727e;
+  --main-white: var(--secondary-50);
+  --main-black: #060718;
+
+  /* Layout variables. */
+  --navbar-height: 84px;
+  --banner-height: 42px;
+  --navbar-height-with-banner: calc(
+    var(--navbar-height) + var(--banner-height)
+  );
 }
