+++ src/components/layout/navbar/app-navbar.css
@@ -1,0 +1,128 @@
+/*
+Navigation bar.
+It is made up of three sections.
+- Left: Logo
+- Middle: Links with animated underline
+- Right: Dropdown link that displays extended section
+
+===================================================================
+| ( LOGO ) |          Home    About    Contact          [v] Expand |
+===================================================================
+
+NOTE: The navbar inherits it's background (not explicitly set) because it is designed to be transparent. If you want a solid background, remove the transparency in the React component styles and add a background color.
+*/
+.navbar-container {
+  position: fixed;
+  top: 0;
+  z-index: 100;
+  border-bottom: 1px solid var(--neutral-300);
+  width: 100%;
+  height: var(--navbar-height);
+
+  @media screen and (min-width: 992px) {
+    border: none;
+  }
+}
+
+.navbar-toolbar-content {
+  margin: auto;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 8px;
+  max-width: calc(var(--screen-xxl) + 72px);
+  height: 100%;
+
+  @media (min-width: 992px) {
+    padding: 0 24px;
+  }
+}
+
+.navbar-link-list-item {
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  font-weight: normal;
+}
+
+.navbar-link {
+  z-index: 15;
+  padding: 1rem;
+  margin:auto;
+}
+
+.navbar-link-selected {
+  font-weight: bold;
+}
+
+.navbar-left-section {
+  display: inline-flex;
+  align-items: center;
+  padding: 1rem;
+  height: 100%;
+}
+
+.navbar-middle-section {
+  display: none;
+  position: relative;
+  column-gap: 50px;
+  justify-content: center;
+  list-style-type: none;
+
+  @media screen and (min-width: 992px) {
+    display: inline-flex;
+  }
+}
+
+/* Sibling selectors to position the underline */
+.one ~ .navbar-underline {
+  display: block;
+  margin-left: -82%;
+}
+
+.two ~ .navbar-underline {
+  display: block;
+  margin-left: -27%;
+}
+
+.three ~ .navbar-underline {
+  display: block;
+  margin-left: 32%;
+}
+
+.navbar-underline {
+  display: none;
+  position: absolute;
+  top: 82%;
+  float: left;
+  z-index: 10;
+  transition: 0.3s ease-in-out;
+  margin: 0;
+  border: none;
+  background-color: var(--neutral-900);
+  width: 5rem;
+  height: 2px;
+}
+
+.navbar-toggle-button {
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  padding: 1rem;
+  color: var(--main-vivid-blue);
+
+  /* Hide potential text inside the button */
+  * {
+    display: none;
+  }
+}
+
+.navbar-toggle-button--collapsed {
+  background: url("/img/navbar/chevron_down.svg") no-repeat center right;
+  padding-right: 30px;
+}
+.navbar-toggle-button--extended {
+  background: url("/img/navbar/chevron_down.svg") no-repeat center right;
+  padding-right: 30px;
+}
