--- src/styles/main.css
@@ -1,71 +1,0 @@
-@import "./button.css";
-@import "./layout.css";
-@import "./typography.css";
-
-:root {
-  /* Application color theme. */
-  --neutral-50: #f7f7f8;
-  --neutral-100: #eaeaec;
-  --neutral-200: #d4d4d8;
-  --neutral-300: #bfbfc5;
-  --neutral-400: #a9aab1;
-  --neutral-500: #72727d;
-  --neutral-600: #55555e;
-  --neutral-700: #3a3a40;
-  --neutral-800: #1d1d20;
-  --neutral-900: #060718;
-  --primary-50: #f0f4fe;
-  --primary-100: #d6e1fe;
-  --primary-200: #adc4fc;
-  --primary-300: #85a6fb;
-  --primary-400: #5b89fa;
-  --primary-500: #0d2866;
-  --primary-600: #0034ae;
-  --primary-700: #002377;
-  --primary-800: #00113c;
-  --primary-900: #00091e;
-  --secondary-50: #fffcf5;
-  --secondary-100: #fff9e7;
-  --secondary-200: #fff2cf;
-  --secondary-300: #ffe5a1;
-  --secondary-400: #ffd874;
-  --secondary-500: #ffbf32;
-  --secondary-600: #ecaf2c;
-  --secondary-700: #e9ad2e;
-  --secondary-800: #251e0d;
-  --secondary-900: #0c0900;
-  --shades-black: #000000;
-  --shades-white: #ffffff;
-
-  /* Brand colors. */
-  --landing-background-color: var(--secondary-50);
-  --navbar-color: var(--seconday-50);
-  --main-vivid-blue: #00478e;
-  --main-light-blue: #00a6fb;
-  --main-grey: #72727e;
-  --main-white: var(--secondary-50);
-  --main-black: #060718;
-
-  /* Screen breakpoints. */
-  --screen-xs: 400px;
-  --screen-sm: 576px;
-  --screen-md: 768px;
-  --screen-lg: 992px;
-  --screen-xl: 1200px;
-  --screen-xxl: 1400px;
-
-  /* Layout variables. */
-  --navbar-height: 84px;
-  --banner-height: 42px;
-  --navbar-height-with-banner: calc(
-    var(--navbar-height) + var(--banner-height)
-  );
-}
-
-h1,
-h2,
-h3,
-h4,
-h5 {
-  text-wrap: pretty;
-}
